welcomeの画面を消してコードで初回起動時の処理を分岐した
